# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FitHub API",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/plans": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "List Plans",
        "operationId": "list_plans_api_v1_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanSummary"
                  },
                  "type": "array",
                  "title": "Response List Plans Api V1 Plans Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Create Plan",
        "operationId": "create_plan_api_v1_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/plans/tasks/{task_id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Task",
        "operationId": "get_task_api_v1_plans_tasks__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Get Plan",
        "operationId": "get_plan_api_v1_plans__plan_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}/today": {
      "get": {
        "tags": [
          "plans"
        ],
        "summary": "Today Session",
        "operationId": "today_session_api_v1_plans__plan_id__today_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/PlannedSessionOut"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "title": "Response Today Session Api V1 Plans  Plan Id  Today Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{plan_id}/revise": {
      "post": {
        "tags": [
          "plans"
        ],
        "summary": "Revise Plan",
        "operationId": "revise_plan_api_v1_plans__plan_id__revise_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanRevisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreatePlanRequest": {
        "properties": {
          "goal": {
            "type": "string",
            "enum": [
              "general_fitness",
              "strength",
              "endurance",
              "competition_prep"
            ],
            "title": "Goal"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "title": "Title"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "weeks": {
            "type": "integer",
            "maximum": 24.0,
            "minimum": 4.0,
            "title": "Weeks"
          },
          "training_age": {
            "type": "string",
            "enum": [
              "beginner",
              "intermediate",
              "advanced"
            ],
            "title": "Training Age"
          }
        },
        "type": "object",
        "required": [
          "goal",
          "title",
          "start_date",
          "weeks",
          "training_age"
        ],
        "title": "CreatePlanRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MesocycleOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phase": {
            "type": "string",
            "title": "Phase"
          },
          "week_start": {
            "type": "integer",
            "title": "Week Start"
          },
          "week_end": {
            "type": "integer",
            "title": "Week End"
          },
          "focus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Focus"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phase",
          "week_start",
          "week_end",
          "focus"
        ],
        "title": "MesocycleOut"
      },
...
      "PlanDetail": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "weeks": {
            "type": "integer",
            "title": "Weeks"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "training_age": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Training Age"
          },
          "mesocycles": {
            "items": {
              "$ref": "#/components/schemas/MesocycleOut"
            },
            "type": "array",
            "title": "Mesocycles"
          },
          "sessions": {
            "items": {
              "$ref": "#/components/schemas/PlannedSessionOut"
            },
            "type": "array",
            "title": "Sessions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "goal",
          "title",
          "branch_name",
          "weeks",
          "status",
          "start_date",
          "end_date",
          "created_at",
          "training_age",
          "mesocycles",
          "sessions"
        ],
        "title": "PlanDetail"
      },
      "PlanRevisionRequest": {
        "properties": {
          "feedback": {
            "type": "string",
            "maxLength": 500,
            "minLength": 5,
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "feedback"
        ],
        "title": "PlanRevisionRequest"
      },
      "PlanSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "goal": {
            "type": "string",
            "title": "Goal"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "branch_name": {
            "type": "string",
            "title": "Branch Name"
          },
          "weeks": {
            "type": "integer",
            "title": "Weeks"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "start_date": {
            "type": "string",
            "format": "date",
            "title": "Start Date"
          },
          "end_date": {
            "type": "string",
            "format": "date",
            "title": "End Date"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "goal",
          "title",
          "branch_name",
          "weeks",
          "status",
          "start_date",
          "end_date",
          "created_at"
        ],
        "title": "PlanSummary"
      },
      "PlanTaskResponse": {
        "properties": {
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "PlanTaskResponse"
      },
      "PlannedItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "movement_name": {
            "type": "string",
            "title": "Movement Name"
          },
          "sets": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sets"
          },
          "reps": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reps"
          },
          "load_pct_1rm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Pct 1Rm"
          },
          "load_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Load Kg"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "item_order": {
            "type": "integer",
            "title": "Item Order"
          }
        },
        "type": "object",
        "required": [
          "id",
          "movement_name",
          "sets",
          "reps",
          "load_pct_1rm",
          "load_kg",
          "notes",
          "item_order"
        ],
        "title": "PlannedItemOut"
      },
      "PlannedSessionOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "mesocycle_id": {
            "type": "string",
            "format": "uuid",
            "title": "Mesocycle Id"
          },
          "scheduled_date": {
            "type": "string",
            "format": "date",
            "title": "Scheduled Date"
          },
          "session_type": {
            "type": "string",
            "title": "Session Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/PlannedItemOut"
            },
            "type": "array",
            "title": "Items",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "mesocycle_id",
          "scheduled_date",
          "session_type",
          "title",
          "notes",
          "status"
        ],
        "title": "PlannedSessionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}